changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11798 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Control/RemoteRunControl/ORRemoteRunModel.h
@@ -139,7 +139,7 @@
 
 #pragma mark ***Delegate Methods
 - (void) netsocketConnected:(id)aSocket;
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount;
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount;
 - (void) netsocketDisconnected:(NetSocket*)inNetSocket;
 
 - (void) connectSocket:(BOOL)state;
